Refactor equals and hashCode of Film
The combination of title and release date used before is not unique and can't
serve as a business key. Due to lack of a proper business key, the new equals
compares by reference or id if it is present for the both entities being
compared. Hashcode returns the same value so its value doesn't depend on entity
state.